Initial Checking
The purpose of this prototype is to demonstrate the use of the binomial
test strategy in order to handle event generation for response times.
We explicitly assume that all tuples can be categorized binomially based
on whether or not the response time is greater than or equal to the
threshold.  We collect tuples into samples using a sliding window that
is contrained based on a min sample size (default is 3), a min
evaluation interval (default is 1 minute), a max queue size (default is
50), and a max time window (default is 3 hours).  We then evaluate this
sample using the binomial test with a configurable expected error rate
(default is 0.10) and an alpha of 0.05.