Added an alpha cooldown algorithm in order to ensure that, if after raising an alert, the response time hovers perpetually between the alerting significance threshold and the dealerting significance threshold, we will eventually issue a RETURN_TO_NORMAL message.  The algorirthm works as follows:
0) Two significance thresholds are defined, one for alerting and one for
dealerting.  The significance threshold for dealerting is greater than
the significance threshold for alerting.

1) The state is initialized to 'normal' and the significance threshold
is initialized to the alerting alpha.

2) If the state is currently 'normal', and the result of the evaluation
is less than the current significance threshold, we switch the state to
'alert', and the significance threshold to the dealerting alpha.

3) If the status is currently 'alert', and the result of the evaluation
is greater than or equal to the current significance threshold, we
switch the state to 'normal', and the significance threshold to the
alerting threshold. 

4) If the status is currently 'alert' and the result of the evaluation
is greater than or equal to the alerting alpha, but less than the
current significance threshold, we decrement the current significance
threshold by a predefined alpha cooldown interval until the significance
threshold converges with the alerting alpha.